Move its Java package to org.embulk.output.gcs from org.embulk.output
