+++ src/main/java/org/embulk/output/gcs/AuthMethod.java
@@ -1,4 +1,4 @@
-package org.embulk.output;
+package org.embulk.output.gcs;
 
 public enum AuthMethod
 {
